(style) 공항/out 모달 구현 및 font사이즈 수정
