+++ src/components/SchedulePage/Changing/OutAirport/OutTraffic.jsx
@@ -7,6 +7,15 @@ import PublicBusIcon from "../../../../img/PublicBus.svg";
 import AirportTransferIcon from "../../../../img/AirportTransfer.svg";
 import "../../../../font/font.css";
 
+import TrainB from "../../../../img/train_black.png";
+import TaxiB from "../../../../img/taxi_black.png";
+import ShuttleB from "../../../../img/shuttle_black.png";
+import PublicB from "../../../../img/public_black.png";
+import TransferB from "../../../../img/transfer_black.png";
+
+import { useState, useEffect } from "react";
+import Schedule_Modal from "../../NoChanging/Schedule_Modal/Schedule_Modal";
+
 const StyleContainer = styled.div`
   background-color: black;
   color: white;
@@ -62,6 +71,39 @@ const PublicBus = styled.img``;
 const AirportTransfer = styled.img``;
 
 export default function OutTraffic() {
+  const viaAirport = localStorage.getItem("viaAirport");
+  const arrivalTime = localStorage.getItem("arrivalTime");
+  const durationTime = localStorage.getItem("durationTime");
+
+  const [infoAirport, setAirport] = useState({});
+
+  /*흰색*/
+  const TrainI = TrainIcon;
+  const TaxiI = TaxiIcon;
+  const ShuttleI = ShuttleIcon;
+  const publicI = PublicBusIcon;
+  const transferI = AirportTransferIcon;
+
+  /*검정*/
+  const TrainIconB = TrainB;
+  const TaxiIconB = TaxiB;
+  const ShuttleIconB = ShuttleB;
+  const publicIconB = PublicB;
+  const transferIconB = TransferB;
+
+  function getAirport() {
+    const url = `https://api.zionhann.shop/app/viaflight/layover-airport?layoverAirportName=${viaAirport}&layoverArrivalTime=${arrivalTime}&layoverTime=${durationTime}`;
+    fetch(url)
+      .then((response) => response.json())
+      .then((data) => {
+        setAirport(data.layoverAirport);
+      });
+  }
+
+  useEffect(() => {
+    getAirport();
+  }, []);
+
   return (
     <StyleContainer>
       <AboutFadin>
@@ -70,23 +112,52 @@ export default function OutTraffic() {
             <Title>공항 - 도심 간 이동 교통편</Title>
             <Right>
               <Card>
-                <Train src={TrainIcon} />
+                <Schedule_Modal
+                  tittle={infoAirport.trainMrtOutAirportHeader}
+                  description={infoAirport.trainMrtOutAirport}
+                  Modalicon={TrainI}
+                  ModaliconBlack={TrainIconB}
+                ></Schedule_Modal>
                 <div>Train MRT</div>
               </Card>
+
               <Card>
-                <Taxi src={TaxiIcon} />
+                <Schedule_Modal
+                  tittle={infoAirport.taxiOutAirportHeader}
+                  description={infoAirport.taxiOutAirport}
+                  Modalicon={TaxiI}
+                  ModaliconBlack={TaxiIconB}
+                ></Schedule_Modal>
                 <div>Taxi</div>
               </Card>
+
               <Card>
-                <Shuttle src={ShuttleIcon} />
+                <Schedule_Modal
+                  tittle={infoAirport.shuttleBusOutAirportHeader}
+                  description={infoAirport.airportShuttleBusOutAirport}
+                  Modalicon={ShuttleI}
+                  ModaliconBlack={ShuttleIconB}
+                ></Schedule_Modal>
                 <div style={{ letterSpacing: "-2px" }}>Shuttle Service</div>
               </Card>
+
               <Card>
-                <PublicBus src={PublicBusIcon} />
+                <Schedule_Modal
+                  tittle={infoAirport.publicBusOutAirportHeader}
+                  description={infoAirport.publicBusOutAirport}
+                  Modalicon={publicI}
+                  ModaliconBlack={publicIconB}
+                ></Schedule_Modal>
                 <div>Public Bus</div>
               </Card>
+
               <Card>
-                <AirportTransfer src={AirportTransferIcon} />
+                <Schedule_Modal
+                  tittle={infoAirport.airportTransferOutAirportHeader}
+                  description={infoAirport.airportTransferBusOutAirport}
+                  Modalicon={transferI}
+                  ModaliconBlack={transferIconB}
+                ></Schedule_Modal>
                 <div style={{ letterSpacing: "-3px" }}>Airport Transfer</div>
               </Card>
             </Right>
